refactor: Adds modularization to plugins directory.
Now it has one file per plugin and it is so much better to manage.
