+++ lua/plugins/cmp-buffer.lua
@@ -1,0 +1,5 @@
+-- cmp-buffer: Buffer text completion source
+return {
+  "hrsh7th/cmp-buffer",
+}
+
