+++ lua/plugins/cmp-luasnip.lua
@@ -1,0 +1,5 @@
+-- cmp_luasnip: LuaSnip completion source
+return {
+  "saadparwaiz1/cmp_luasnip",
+}
+
