+++ lua/plugins/cmp-nvim-lsp.lua
@@ -1,0 +1,5 @@
+-- cmp-nvim-lsp: LSP completion source
+return {
+  "hrsh7th/cmp-nvim-lsp",
+}
+
