+++ lua/plugins/cmp-path.lua
@@ -1,0 +1,5 @@
+-- cmp-path: File path completion source
+return {
+  "hrsh7th/cmp-path",
+}
+
