+++ lua/plugins/friendly-snippets.lua
@@ -1,0 +1,5 @@
+-- friendly-snippets: Collection of useful snippets
+return {
+  "rafamadriz/friendly-snippets",
+}
+
