+++ lua/plugins/luasnip.lua
@@ -1,0 +1,13 @@
+-- LuaSnip: Snippet engine
+return {
+  "L3MON4D3/LuaSnip",
+  version = "v2.*",
+  build = "make install_jsregexp",
+  dependencies = {
+    "rafamadriz/friendly-snippets",
+  },
+  config = function()
+    require("luasnip.loaders.from_vscode").lazy_load()
+  end,
+}
+
