+++ lua/plugins/telescope-fzf-native.lua
@@ -1,0 +1,7 @@
+-- telescope-fzf-native: FZF sorter for telescope (better performance)
+return {
+  "nvim-telescope/telescope-fzf-native.nvim",
+  build = "make",
+  lazy = true,
+}
+
