+++ lua/plugins/vim-jsx-pretty.lua
@@ -1,0 +1,9 @@
+-- vim-jsx-pretty: Enhanced JSX/HTML syntax highlighting
+return {
+  "MaxMEllon/vim-jsx-pretty",
+  lazy = false,
+  config = function()
+    vim.g.jsx_ext_required = 0
+  end,
+}
+
