+++ version.lua
@@ -1,6 +1,6 @@
 return {
-  version = "1.0.0",
-  build_date = "2025-10-28",
+  version = "1.1.0",
+  build_date = "2025-12-15",
   neovim_min_version = "0.11.0",
   description = "AmetAlvirde's personal dotNvim from scratch configuration"
 }
